Added TWDH_MODE variable
TWDH_MODE should be set in docker/.env-ckan.env

TWDH_MODE is only relevant if the Docker images were build using the
`build-dev` target.

If TWDH_MODE is not set, the default is  TWDH_MODE = dev

When TWDH_MODE = debug, CKAN is started with the python debugpy module
enabled, allowing a developer to attach the VSCode debugger. However, in
this mode, CKAN does not restart when file changes are synced into the
docker container. This means that the developer has to exec into the
CKAN pod and manually `kill -HUP [ckan pid]` if they need the CKAN
server to pick up code changes.

When TWDH_MODE = dev, ckan is started in the usual way with UWSGI with
CKAN config variable `debug = true` set.