WIP Updated overwrite.schema.xml SOLR schema URL
We need to use the customized TWDH version so we are no longer pulling
from CKAN core.

Currently it is pointing to a version in the `twdh-0.8.0-dev` branch.
This needs to be changed once the branch gets merged into main.

It would be better to do this by mounting the file in docker-compose,
but when I tried that there was, I think, a kind of race condition
where the schema file was being used before the JTS jar was in place
and it was causing a crash. Or maybe something else, but I was
getting a SOLR JTS not found exception.