Fixed docker 'sync' action to work with ckan core
Docker 'sync' was syncing all of the extensions correctly but not the
ckan core tree.

Turns out it is because on our development instances the directory
structure is like so:

src/ckan/...
src/ckanext-agsview/...
src/ckanext-twdh_theme/...
...

But inside the container the src tree is set up like so:

src/ckan/...
src/ckan/ckanext-agsview/...
src/ckan/ckanext-twdh_theme/...
...

That is, all of the extensions are inside the ckan src tree instead of
parallel to it. We cannot change the structure on our development
instances without causing git havoc, and I'm not sure why the container
tree is set up the way it is, so I have made the verbose changes seen in
this patch such that each extension and the core tree all get their own
sync blocks.

The sync we had before was correct for the extentions but was silently
failing for the ckan src tree. Worse, docker watch was reporting the
ckan src tree syncs as if they were working, but they not.

This is what the sync statement looked like before.

'''
develop:
  watch:
    - action: sync
      path: /usr/lib/ckan/default/src
      target: /srv/app/src/ckan
'''

You can see the changes I made in this patch to fix the problem.